Replace check for Advanced SIMD
Although `hw.optional.AdvSIMD` is the replacement for `hw.optional.neon`, this parameter is not always present in different versions of
the OS. This may lead to the test suite crashing or tests being erroneously skipped.

Instead, we check if the machine supports `hw.optional.arm64` and, if
true, we can assume Advanced SIMD support is always present.

Signed-off-by: Jakub Sujak <jakub.sujak@arm.com>

Approved-by: Viet-Hoa Do <viet-hoa.do@arm.com>