Reduce noise from comparison
This change bundles some cleanup done while doing FP16 IGEMM work. The
main change is the reduced output from the comparison function, which
instead of printing each differing value on a new line instead bundles
them on a per matrix row, with some block description. Also, the output
is only printed if there is a mismatch

Signed-off-by: Emil Ohlsson <emil.ohlsson@arm.com>

Approved-by: Jakub Sujak <jakub.sujak@arm.com>