Allow matmul_clamp_f32_qai8dxp_qsi4c32p tests to run without bfloat16 CPU support
Related to #9 - allows some tests to run when the CPU doesn't have bfloat16 support, by reducing need for bfloat16 support in the test framework.

Use of bfcvt in the test framework is replaced by simple truncation - this will effectively
round towards 0.

Add a few tests for negative numbers in the test framework BFloat16 code, and allow running the
unit tests for the framework code on CPUs without bfloat16 hardware support.

Signed-off-by: Matthew Bentham <Matthew.Bentham@arm.com>

Reviewed-by: Viet-Hoa Do <viet-hoa.do@arm.com>
Approved-by: Anton Bondarenko <anton.bondarenko@arm.com>